+++ footlog/src/app/onboarding/step4/page.tsx
@@ -1,14 +1,13 @@
 'use client';
 import { useEffect, useState } from 'react';
-import { useRecoilValue } from 'recoil';
+import { useRouter } from 'next/navigation';
 import { CheckIcon } from '@public/icon';
 import OnboardingTitle from '@components/onboarding/OnboardingTitle';
 import OnboardingBtn from '@components/onboarding/OnboardingBtn';
 import { MoonLoader } from 'react-spinners';
-import { firstOnboardingState, secondOnboardingState, thirdOnboardingState } from '@recoil/atom';
-import usePostPreferKeyword from '@hooks/onboarding/usePostPreferKeyword';
 
 export default function page() {
+  const router = useRouter();
   const [isOnboardingBtnDisabled, setIsOnboardingBtnDisabled] = useState(true); // 초기값 true로 설정
   const [titleText, setTitleText] = useState(
     <>
@@ -17,11 +16,6 @@ export default function page() {
       플로깅 코스를 찾고 있어요.
     </>,
   );
-
-  const firstState = useRecoilValue(firstOnboardingState);
-  const secondState = useRecoilValue(secondOnboardingState);
-  const thirdState = useRecoilValue(thirdOnboardingState);
-  const { mutate: postPreferKeywordMutate } = usePostPreferKeyword();
 
   // 일단 3초 후 바뀌게끔 해놨는데 나중에 api 연결로 바꿀게!
   useEffect(() => {
@@ -40,14 +34,6 @@ export default function page() {
     return () => clearTimeout(timer); // 컴포넌트 언마운트 시 타이머 정리
   }, []);
 
-  function handleOnboardingBtn() {
-    postPreferKeywordMutate({
-      firstOnboardingState: firstState,
-      secondOnboardingState: secondState,
-      thirdOnboardingState: thirdState,
-    });
-  }
-
   return (
     <main className="relative flex h-full w-full flex-col px-24pxr pt-100pxr">
       <OnboardingTitle text={titleText} />
@@ -58,7 +44,11 @@ export default function page() {
           <CheckIcon /> // 버튼이 활성화되면 CheckIcon 출력
         )}
       </div>
-      <OnboardingBtn text="시작하기" $disabled={isOnboardingBtnDisabled} handleOnboardingBtn={handleOnboardingBtn} />
+      <OnboardingBtn
+        text="시작하기"
+        $disabled={isOnboardingBtnDisabled}
+        handleOnboardingBtn={() => router.push('/home')}
+      />
     </main>
   );
 }
